Move response stuff to response.ts
